cast to types expected by Win32 API
```
src/agent.c:314:..: warning C4267: '=': conversion from 'size_t' to 'DWORD', possible loss of data
src/agent.c:314:18: warning: conversion from 'size_t' {aka 'long long unsigned int'} to 'DWORD' {aka 'long unsigned int'} may change value [-Wconversion]
src/agent.c:314:20: warning: implicit conversion loses integer precision: 'unsigned long long' to 'DWORD' (aka 'unsigned long') [-Wshorten-64-to-32]
src/wincng.c:671:..: warning C4267: 'function': conversion from 'size_t' to 'ULONG', possible loss of data
src/wincng.c:671:39: warning: conversion from 'size_t' {aka 'long long unsigned int'} to 'ULONG' {aka 'long unsigned int'} may change value [-Wconversion]
src/wincng.c:671:39: warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long long') to 'ULONG' (aka 'unsigned long') [-Wshorten-64-to-32]
src/wincng.c:2552:33: warning: conversion from 'long long unsigned int' to 'DWORD' {aka 'long unsigned int'} may change value [-Warith-conversion]
```

+++ src/wincng.c
@@ -668,7 +668,7 @@ _libssh2_wincng_key_sha_verify(_libssh2_wincng_key_ctx *ctx,
     memcpy(data, sig, datalen);
 
     ret = BCryptVerifySignature(ctx->hKey, pPaddingInfo,
-                                hash, hashlen, data, datalen, flags);
+                                hash, (ULONG)hashlen, data, datalen, flags);
 
     _libssh2_wincng_safe_free(hash, hashlen);
     _libssh2_wincng_safe_free(data, datalen);
@@ -2551,7 +2551,8 @@ _libssh2_dh_secret(_libssh2_dh_ctx *dhctx, _libssh2_bn *secret,
         unsigned char *start, *end;
         BCRYPT_DH_KEY_BLOB *public_blob = NULL;
         DWORD key_length_bytes = max(f->length, dhctx->dh_params->cbKeyLength);
-        DWORD public_blob_len = sizeof(*public_blob) + 3 * key_length_bytes;
+        DWORD public_blob_len = (DWORD)(sizeof(*public_blob) +
+                                        3 * key_length_bytes);
 
         {
             /* Populate a BCRYPT_DH_KEY_BLOB; after the header follows the
